chore: split up into different files
`ghafs.go` now focuses solely on implementing the interface to make up
the filesystem.
