+++ ghafs.go
@@ -12,44 +12,21 @@ import (
 	"github.com/google/go-github/v28/github"
 )
 
-type ReleaseAssets struct {
-	release *github.RepositoryRelease
-	assets  []*github.ReleaseAsset
-}
-
-type ReleaseAssetsMap map[string]ReleaseAssets
-
-func rasToDirents(rasm *ReleaseAssetsMap) []fuse.Dirent {
-	dirents := make([]fuse.Dirent, len(*rasm))
-
-	i := 0
-	for tag, ras := range *rasm {
-		dirents[i] = fuse.Dirent{
-			Inode: uint64(ras.release.GetID()),
-			Name:  tag,
-			Type:  fuse.DT_Dir,
-		}
-		i++
-	}
-	return dirents
-}
-
-func assetsToDirents(assets []*github.ReleaseAsset) []fuse.Dirent {
-	dirents := make([]fuse.Dirent, len(assets))
-
-	for i, asset := range assets {
-		dirents[i] = fuse.Dirent{
-			Inode: uint64(asset.GetID()),
-			Name:  asset.GetName(),
-			Type:  fuse.DT_File,
-		}
-	}
-	return dirents
-}
-
 // ghaFS implements the GitHub Release Assets file system.
 type ghaFS struct {
 	rasm  *ReleaseAssetsMap
+	token *string
+}
+
+// root implements both Node and Handle for the root directory.
+type root struct {
+	rasm  *ReleaseAssetsMap
+	token *string
+}
+
+// tagDir implements both Node and Handle for the root directory.
+type tagDir struct {
+	ras   *ReleaseAssets
 	token *string
 }
 
@@ -61,16 +38,14 @@ func (g ghaFS) Root() (fs.Node, error) {
 	return root{rasm: g.rasm, token: g.token}, nil
 }
 
-// root implements both Node and Handle for the root directory.
-type root struct {
-	rasm  *ReleaseAssetsMap
-	token *string
-}
-
 func (root) Attr(ctx context.Context, a *fuse.Attr) error {
 	a.Inode = 1
 	a.Mode = os.ModeDir | 0555
 	return nil
+}
+
+func (r root) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
+	return rasToDirents(r.rasm), nil
 }
 
 func (r root) Lookup(ctx context.Context, name string) (fs.Node, error) {
@@ -83,20 +58,14 @@ func (r root) Lookup(ctx context.Context, name string) (fs.Node, error) {
 	return nil, fuse.ENOENT
 }
 
-func (r root) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
-	return rasToDirents(r.rasm), nil
-}
-
-// tagDir implements both Node and Handle for the root directory.
-type tagDir struct {
-	ras   *ReleaseAssets
-	token *string
-}
-
 func (t tagDir) Attr(ctx context.Context, a *fuse.Attr) error {
 	a.Inode = uint64(t.ras.release.GetID())
 	a.Mode = os.ModeDir | 0555
 	return nil
+}
+
+func (t tagDir) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
+	return assetsToDirents(t.ras.assets), nil
 }
 
 func (t tagDir) Lookup(ctx context.Context, name string) (fs.Node, error) {
@@ -106,10 +75,6 @@ func (t tagDir) Lookup(ctx context.Context, name string) (fs.Node, error) {
 		}
 	}
 	return nil, fuse.ENOENT
-}
-
-func (t tagDir) ReadDirAll(ctx context.Context) ([]fuse.Dirent, error) {
-	return assetsToDirents(t.ras.assets), nil
 }
 
 // assetFile implements both Node and Handle for the hello file.
