+++ ghafsMapping.go
@@ -1,0 +1,34 @@
+package main
+
+import (
+	"bazil.org/fuse"
+	"github.com/google/go-github/v28/github"
+)
+
+func rasToDirents(rasm *ReleaseAssetsMap) []fuse.Dirent {
+	dirents := make([]fuse.Dirent, len(*rasm))
+
+	i := 0
+	for tag, ras := range *rasm {
+		dirents[i] = fuse.Dirent{
+			Inode: uint64(ras.release.GetID()),
+			Name:  tag,
+			Type:  fuse.DT_Dir,
+		}
+		i++
+	}
+	return dirents
+}
+
+func assetsToDirents(assets []*github.ReleaseAsset) []fuse.Dirent {
+	dirents := make([]fuse.Dirent, len(assets))
+
+	for i, asset := range assets {
+		dirents[i] = fuse.Dirent{
+			Inode: uint64(asset.GetID()),
+			Name:  asset.GetName(),
+			Type:  fuse.DT_File,
+		}
+	}
+	return dirents
+}
